+++ mex/common/primary_source/helpers.py
@@ -8,26 +8,27 @@ from mex.common.primary_source.transform import (
     transform_seed_primary_sources_to_extracted_primary_sources,
 )
 
-### this code is nearly the same as in the pytest.fixture.extracted_primary_sources.
-### Can we reuse this?
+# this code is nearly the same as in the pytest.fixture.extracted_primary_sources.
+# Can we reuse this?
 
 
 @cache
-def get_all_extracted_primary_sources() -> list[ExtractedPrimarySource]:
+def get_all_extracted_primary_sources() -> dict[str, ExtractedPrimarySource]:
     """Extract and transform all primary sources.
 
     Extract the primary sources from the raw-data JSON file and transform them into
-    a list of ExtractedPrimarySources.
+    a dictionary of ExtractedPrimarySources.
 
     Returns:
-        List of all ExtractedPrimarySources
+        dictionary of all ExtractedPrimarySources
     """
     seed_primary_sources = extract_seed_primary_sources()
-    return list(
+    extracted_primary_sources = (
         transform_seed_primary_sources_to_extracted_primary_sources(
             seed_primary_sources
         )
     )
+    return {p.identifierInPrimarySource: p for p in extracted_primary_sources}
 
 
 @cache
@@ -40,7 +41,9 @@ def get_extracted_primary_source_by_name(name: str) -> ExtractedPrimarySource | 
     Returns:
         Extracted primary source if it was found, else None
     """
-    primary_sources_by_name = {
-        p.identifierInPrimarySource: p for p in get_all_extracted_primary_sources()
-    }
-    return primary_sources_by_name.get(name)
+    try:
+        extracted_primary_source = get_all_extracted_primary_sources()[name]
+    except KeyError:
+        return None
+
+    return extracted_primary_source