+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - make instances of extracted data hashable
 - BREAKING: Wikidata convenience function refactored and renamed to 'helper'
 - wikidata helper function split between mex-common and mex-extractors
+- code de-duplication: fixture extracted_primary_sources now uses function-part of primary sources helper
 
 ### Deprecated